+++ backend/games/.flake8
@@ -1,4 +1,4 @@
 [flake8]
 exclude = venv, migrations, __init__.py
 max-line-length=100
-ignore=F401
+ignore=F401,W503
